spex

Tools for spatial extents that are agnostic regarding format (i.e. sp, sf, or raster). These functions fill some of the gaps within and between these packages for dealing with object extents in flexible ways. Generally, spex considers extents of raster cells, and extents of objects as first-class objects (with projection metadata). and provides helpers for latitudinal boundaries within projected data.

Spatial Extent

buffer_extent

Buffer an extent to a given whole number.

latitudecircle

Create a latitude circle in a chosen projection.

latmask

Mask a raster based on a minimum (or maximum) latitude.

polygonize

Convert raster cells to polygons.

qm_rasterToPolygons

The

sf	version of

polygonize	.

qm_rasterToPolygons_sp

The

sp	version of

polygonize	.

spex

A function to produce a fully fledged Spatial object extent.

xlim, ylim

Helper functions for extents.